Updated files

+++ api/authors/delete.php
@@ -1,39 +1,45 @@
 <?php
 
 // accessing our front-facing API
-// Header notic how the heder names are in capital letters 
 header("Access-Control-Allow-Origin: *");
 header("Content-Type: application/json");
 header("Access-Control-Allow-Methods: DELETE");
 header("Access-Control-Allow-Headers: Access-Control-Allow-Headers, Content-Type, Access-Control-Allow-Methods, Authorization, X-Requested-With");
-// the above header values all must be on ONE LINE 
-
 
 include_once '../../config/Database.php';
 include_once '../../models/Author.php';
 
 // Instantiate/Create DB and Connect
-$database = new Database(); // new database object
-$db = $database->connect(); // the connect pre-defined function
+$database = new Database(); 
+
+// new database object
+$db = $database->connect(); 
 
 
-// Instantiate/create new (blog) post object
+// Instantiate/create new author object
 $author = new Author($db);
 
 
-// Gt the raw posted data 
 $data = json_decode(file_get_contents("php://input"));
 
 
 
-if ($author->delete())
-{
-    // encode to JSON
-    echo json_encode(array("message" => "Author Deleted"));
+ $author->id = isset($data->id) ? $data->id : null;
 
+
+if (isset($author->id))
+{ 
+        
+    if ($author->delete())
+    {
+        // encode to JSON
+        echo json_encode(value: array("message" => "Author Deleted"));
+    }
+    
 }
 else
 {
 
-    echo json_encode(array("message" => "Author Not Deleted"));
+    echo json_encode(value: array("message" => "Author Not Deleted"));
+
 }
